requirements.txt
pandas==1.3.0
numpy>=1.18.5
seaborn==0.10.1
corpy==0.3.1
torch==1.13.1
torchaudio==0.13.1
torchtext==0.6.0
scikit-learn==0.22.2
fasttext-wheel==0.9.2
gensim==3.8.3
Cython==0.29.14
deprecation==2.1.0
matplotlib
tqdm==4.61.2
nltk==3.5


